requirements.txt
amqp==2.5.2
# anyjson==0.3.3
billiard==3.6.1.0
blessings==1.7
bpython==0.18
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
curtsies==0.3.1
Django==2.2.13
django-admin==2.0.0
django-debug-toolbar==2.1
django-excel-response2==2.0.8
django-extensions==2.2.6
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-render-block==0.6
django-sendgrid-v5==0.8.1
django-six==1.0.4
django-templated-email==2.3.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.18.2
greenlet==0.4.15
idna==2.8
image==1.5.28
importlib-metadata==1.5.0
jdcal==1.4.1
kombu==4.6.7
#mysqlclient==1.4.6
openpyxl==3.0.3
Pillow==7.0.0
Pygments==2.5.2
python-http-client==3.2.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
screen==1.0.1
sendgrid==6.1.1
six==1.14.0
sqlparse==0.3.0
typing==3.7.4.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
xlwt==1.3.0
zipp==2.1.0
